Compiler portability changes, for files in booster folder: 1. In booster.c, don't include <getopt.h> under MSVC; and 2. define _CRT_SECURE_NO_WARNINGS under MSVC     (also done in tree.c); and 3. use _strdup rather strdup (under MSVC). 4. In stats.c, replace on-stack dynamic arrays with calls     to malloc, in median_int_vec() and median_double_vec()     (conditionally, under Microsoft Compiler).     Note that neither of these functions are quite right!     (not when length is even).     That same approach was already used in shuffle(). 5. in tree.c, use duplicate_string macro in place of strdup     or _strdup (definition depends on compiler).

+++ booster/hashmap.c
@@ -191,7 +191,7 @@ unsigned long crc32_booster(const unsigned char *s, unsigned int len)
  */
 unsigned int hashmap_hash_int(hashmap_map * m, char* keystring){
 
-    unsigned long key = crc32_booster((unsigned char*)(keystring), strlen(keystring));
+    unsigned long key = crc32_booster((unsigned char*)(keystring), (unsigned int)strlen(keystring));
 
 	/* Robert Jenkins' 32 bit Mix Function */
 	key += (key << 12);
